Move around all components to their own folders for readability
